% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xin_ctrl2.R
\name{oin_ctrl2}
\alias{oin_ctrl2}
\title{Improve handling of incorrect values to sweidnumbr::oin_ctrl()}
\usage{
oin_ctrl2(x)
}
\arguments{
\item{x}{Vector to check}
}
\value{
Logical vector indicating if a pin is correct (\code{TRUE}) or not (\code{FALSE})
}
\description{
\code{oin_ctrl2} This functions returns FALSE if values isn't a valid pin and returns TRUE if valid.
}

