% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xin_ctrl2.R
\name{pin_ctrl2}
\alias{pin_ctrl2}
\title{Improve handling of incorrect values to sweidnumbr::pin_ctrl()}
\usage{
pin_ctrl2(x)
}
\arguments{
\item{x}{Vector to check}
}
\value{
Logical vector indicating if a pin is correct (\code{TRUE}) or not (\code{FALSE})
}
\description{
\code{pin_ctrl2} This functions returns FALSE if values isn't a valid pin and returns TRUE if valid.
}

